Skip Mujoco tests if OpenGL is not available
When selecting OSMesa by setting the environment variable
`MUJOCO_GL=osmesa`, this does not lead to an `ImportError` from Mujoco,
but to an `AttributeError` from OpenGL, if OSMesa is not available. This
commit adds `AttributeError` to the expected errors, when making the
test environments. It triggers a warning, but does not lead to a skipped
test. This behavior is analog to the case when EGL is selected, but not
available.

Additionally this commit tests if an OffScreenViewer object can be
instantiated. It fails when setting the environment variable
`MUJOCO_GL=glfw`, if GLFW is not available. In this case, the commit
skips the rendering tests with Mujoco environments.