Update multi_binary.py
The sample method in the MultiBinary class currently allows the mask argument to have a shape that is different from the space's shape. This could lead to an error if the mask argument is not the same shape as the space's shape. To fix this, we can add an assertion to the sample method that checks that the mask argument has the same shape as the space's shape.