Update mountain_car.py
from my understanding in the `_height()` method, you are using the `sin()` function to calculate the height of the car. However, the `sin()` function only returns values between -1 and 1. This means that the height of the car could be negative, which would not be physically possible. we should instead use the numpy.clip() function to clamp the values of the sin() function between 0 and 1. am I right?