Update reinforce_invpend_gym_v26.py
In the previous version, we have transitioned to using tensor operations to calculate the loss, but the way the loss is computed currently will give us a tensor of the same shape as deltas, which is not what we want. To fix this and compute the correct total loss, we should use the tensor sum() function to add up all the element-wise products.