Fix scoping of variables
Since the env and agent were defined in the same file,
the BlackjackAgent's referencing to 'env' always worked,
but technically it wasn't in the scope of the agent.

The same happend with reference to epsilon_decay value
which was also initialized in the agents scope.

Specifying the env in the init, reset en get_action
functions makes the separation of the classes proper.