naming returned values from step()
I've added a suggestion to name the values returned from step(), since whenever i try to learn gym I have tons of troubles with this method and it's returned values. The documentation at gymnasium.farama.org is good, but I still manage to get confused and hit walls. Besides, I believe self documenting code is way superior to documentation. I just don't know if there is any penalty to naming the values like I suggest here. If there is no, or if it is a negligible penalty, then will you please consider this for the sake of other head-bangers such as myself?